requirements.txt
flask
python-dotenv
langchain_community
langchain
pypdf
openai
tiktoken
chromadb
